requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
SQLAlchemy==2.0.23
qrcode==7.4.2
Pillow==10.1.0
requests==2.31.0
cryptography==41.0.7
pydantic==2.5.2
aiosqlite==0.19.0
